refactor(mcp): extract cache invalidation helpers to utils.py
Move _invalidate_novel_cache, _invalidate_character_cache,
_invalidate_chapter_cache from novel_tools.py to new utils.py.
Update all importers.
